feat(runtime,mir): cancellation, closures, execution context, and actor dispatch
Add the cancellation-token substrate and executable cancellation scopes, the pipe-closure
surface with binding-accurate captures, and the HewExecutionContext carrier threaded
through dispatch, locks, and spawn. Lower actor receive/init/lifecycle bodies, select { }
over ask and after-timer arms, and the #[on(start)]/#[on(stop)] hooks.

+++ hew-runtime/src/io_time.rs
@@ -95,7 +95,7 @@ pub unsafe extern "C" fn hew_read_file(path: *const c_char) -> *mut c_char {
 ///
 /// No preconditions — delegates to the OS.
 #[no_mangle]
-pub unsafe extern "C" fn hew_sleep_ms(ms: c_int) {
+pub unsafe extern "C" fn hew_sleep_ms(ms: i64) {
     if ms > 0 {
         // SAFETY: ms > 0 checked above, so cast is lossless.
         #[expect(clippy::cast_sign_loss, reason = "guarded by ms > 0")]
